Refactor event detail page into reusable widgets
