+++ lib/features/events/presentation/detail/widgets/event_detail_bottom_bar.dart
@@ -1,0 +1,50 @@
+import 'package:crew_app/l10n/generated/app_localizations.dart';
+import 'package:flutter/material.dart';
+
+class EventDetailBottomBar extends StatelessWidget {
+  final AppLocalizations loc;
+  final VoidCallback onFavorite;
+  final VoidCallback onRegister;
+
+  const EventDetailBottomBar({
+    super.key,
+    required this.loc,
+    required this.onFavorite,
+    required this.onRegister,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      child: Container(
+        padding: const EdgeInsets.fromLTRB(16, 8, 16, 12),
+        decoration: const BoxDecoration(color: Colors.white),
+        child: Row(
+          children: [
+            IconButton(
+              icon: const Icon(Icons.favorite_border),
+              onPressed: onFavorite,
+            ),
+            const SizedBox(width: 12),
+            Expanded(
+              child: SizedBox(
+                height: 48,
+                child: ElevatedButton(
+                  onPressed: onRegister,
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.orange,
+                    foregroundColor: Colors.white,
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                  ),
+                  child: Text(loc.action_register),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
